工作台业务代码导入, 同时调整了部分数据统计模块内容的函数名与返回类型以适配工作台业务代码

+++ sky-server/src/main/java/com/sky/mapper/OrderMapper.java
@@ -79,7 +79,7 @@ public interface OrderMapper {
      * @param map
      * @return
      */
-    Long countByMap(Map map);
+    Integer countByMap(Map map);
 
     /**
      * 统计订单总数
